refactor: rename Analyzer to ComponentProcessor, move to analysis module
