+++ analysis/src/main/java/net/laprun/sustainability/power/analysis/HdrHistogramComponentProcessor.java
@@ -1,15 +1,15 @@
-package net.laprun.sustainability.power.measure;
+package net.laprun.sustainability.power.analysis;
 
 import org.HdrHistogram.IntCountsHistogram;
 
 @SuppressWarnings("unused")
-public class HdrHistogramAnalyzer implements Analyzer {
+public class HdrHistogramComponentProcessor implements ComponentProcessor {
     private static final int HIGHEST_TRACKABLE_VALUE = 1_000_000;
     private static final int NUMBER_OF_SIGNIFICANT_VALUE_DIGITS = 4;
     private static final int CONVERSION_FACTOR = 1000;
     private final IntCountsHistogram histogram;
 
-    public HdrHistogramAnalyzer() {
+    public HdrHistogramComponentProcessor() {
         histogram = new IntCountsHistogram(HIGHEST_TRACKABLE_VALUE, NUMBER_OF_SIGNIFICANT_VALUE_DIGITS);
     }
