+++ measure/src/main/java/net/laprun/sustainability/power/analysis/Analyzers.java
@@ -1,0 +1,4 @@
+package net.laprun.sustainability.power.analysis;
+
+public interface Analyzers {
+}
