+++ measure/src/main/java/net/laprun/sustainability/power/analysis/ComponentProcessor.java
@@ -1,0 +1,6 @@
+package net.laprun.sustainability.power.analysis;
+
+public interface ComponentProcessor {
+    default void recordComponentValue(double value, long timestamp) {
+    }
+}
