--- measure/src/main/java/net/laprun/sustainability/power/measure/Analyzer.java
@@ -1,5 +1,0 @@
-package net.laprun.sustainability.power.measure;
-
-public interface Analyzer {
-    void recordComponentValue(double value, long timestamp);
-}
